requirements.txt
Flask==3.1.0
gunicorn==23.0.0
coqui-tts==0.25.3
pydub==0.25.1
nltk==3.9.1
transformers==4.46.2
torch==2.5.1
pytest==8.3.4
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-flask==1.3.0